Abdm M1 V3 Changes (#52)
* Abdm Facility fetch services changes

* facility save after carecontext and patientcarecontextx update

* FacilityId save in mongo db and Variable declaration correction

* Mongo carecontext save query logic changes

* ABHA Creation M1 V3 API changes

* Abha V3 changes

* removed unused code

---------

+++ src/main/java/com/wipro/fhir/repo/common/AllergyIntoleranceDataModelRepo.java
@@ -21,6 +21,8 @@
 */
 package com.wipro.fhir.repo.common;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RestResource;
 import org.springframework.stereotype.Repository;
@@ -29,7 +31,7 @@ import com.wipro.fhir.data.resource_model.AllergyIntoleranceDataModel;
 
 @Repository
 @RestResource(exported = false)
-public interface AllergyIntoleranceDataModelRepo extends CrudRepository<AllergyIntoleranceDataModel, Long> {
+public interface AllergyIntoleranceDataModelRepo extends CrudRepository<AllergyIntoleranceDataModel, BigInteger> {
 
 
 }
